requirements.txt
fastapi[all]>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
jinja2>=3.1.0
sse-starlette>=2.0.0
snowflake-connector-python>=3.0.0
anthropic>=0.40.0
beautifulsoup4>=4.12.0
